refactor: merge version sorting logic
