+++ src/utils/sort-versions.ts
@@ -2,28 +2,54 @@ import * as semver from 'semver';
 
 import { RunnableVersion } from '../interfaces';
 
-function electronSemVerCompare(a: semver.SemVer, b: semver.SemVer) {
-  const l = a.compareMain(b);
-  if (l) return l;
-  // Electron's approach is nightly -> other prerelease tags -> stable,
-  // so force `nightly` to sort before other prerelease tags.
-  const [prea] = a.prerelease;
-  const [preb] = b.prerelease;
-  if (prea === 'nightly' && preb !== 'nightly') return -1;
-  if (prea !== 'nightly' && preb === 'nightly') return 1;
-  return a.comparePre(b);
-}
+const preTags = ['nightly', 'alpha', 'beta'];
 
-function compare(a: semver.SemVer | string, b: semver.SemVer | string) {
-  // non-semver goes first, sorted lexicographically
-  const astr = typeof a === 'string';
-  const bstr = typeof b === 'string';
-  if (astr && bstr) return (a as string).localeCompare(b as string);
-  if (astr) return -1;
-  if (bstr) return 1;
+/**
+ * Sorts prerelease tags such that nightly -> alpha -> beta.
+ *
+ * @param a Prerelease tag data for the old version.
+ * @param b Prerelease tag data for the new version.
+ * @returns 0 | 1 | -1
+ */
+const preCompare = (
+  a: readonly (string | number)[],
+  b: readonly (string | number)[],
+) => {
+  const one = preTags.indexOf(a[0] as string);
+  const two = preTags.indexOf(b[0] as string);
+  if (one === two) {
+    // Whether the prerelease tag number is the same
+    // e.g. alpha.1 & alpha.1.
+    if (a[1] === b[1]) return 0;
+    return a[1] > b[1] ? 1 : -1;
+  }
 
-  // then semver, sorted newest-to-oldest
-  return -electronSemVerCompare(a as semver.SemVer, b as semver.SemVer);
+  return one > two ? 1 : -1;
+};
+
+/**
+ * Custom semver comparator which takes into account Electron's prerelease
+ * tag hierarchy.
+ *
+ * Sorts in ascending order when passed to Array.sort().
+ *
+ * @param a The old Electron version.
+ * @param b The new Electron version.
+ * @returns 0 | 1 | -1
+ */
+export function semverCompare(a: string, b: string) {
+  const hasPre = (v: string) => preTags.some((tag) => v.includes(tag));
+
+  // Check that major.minor.patch are the same for a and b.
+  if (semver.coerce(a)?.raw === semver.coerce(b)?.raw) {
+    if (hasPre(a) && hasPre(b)) {
+      const { prerelease: aPre } = semver.parse(a) as semver.SemVer;
+      const { prerelease: bPre } = semver.parse(b) as semver.SemVer;
+      return preCompare(aPre, bPre);
+    }
+  }
+
+  return semver.compare(a, b);
 }
 
 /**
@@ -33,14 +59,15 @@ function compare(a: semver.SemVer | string, b: semver.SemVer | string) {
  * @returns {RunnableVersion[]}
  */
 export function sortVersions(versions: RunnableVersion[]): RunnableVersion[] {
-  type VerSemRun = [
-    ver: string,
-    sem: semver.SemVer | null,
-    run: RunnableVersion,
-  ];
-  const sorted = versions
-    .map((run): VerSemRun => [run.version, semver.parse(run.version), run])
-    .sort(([vera, sema], [verb, semb]) => compare(sema || vera, semb || verb));
-  sorted.forEach(([_1, _2, run], idx) => (versions[idx] = run));
-  return versions;
+  type VerSemRun = [ver: string, run: RunnableVersion];
+
+  const filtered = versions.filter((runnable) =>
+    semver.valid(runnable.version),
+  );
+  const sorted = filtered
+    .map((run): VerSemRun => [run.version, run])
+    .sort(([vera], [verb]) => -semverCompare(vera, verb));
+  sorted.forEach(([_1, run], idx) => (filtered[idx] = run));
+
+  return filtered;
 }
